+++ swmmio/definitions.py
