+++ swmmio/reporting/reporting.py
@@ -14,7 +14,7 @@ from swmmio import swmmio
 import os
 import math
 import pandas as pd
-from definitions import *
+from swmmio.definitions import *
 import json, geojson
 import shutil
